--- src/main/java/kono/ceu/gtconsolidate/api/recipe/GTConsolidateRecipeMaps.java
@@ -1,3 +1,0 @@
-package kono.ceu.gtconsolidate.api.recipe;
-
-public class GTConsolidateRecipeMaps {}
